requirements.txt
ttkbootstrap==1.6.1
customtkinter>=5.2,<6
PySide6>=6.6,<7
Pillow==10.1.0
python-dateutil==2.8.2
reportlab==4.0.7
openpyxl==3.1.5
